[Feature] [Balance] Mood Traits & Things (#961)
* Mood Balance

* Update genericNegativeEffects.yml

* fixdashit

* upd

---------
