+++ Resources/Locale/ru-RU/backmen/trait/Mood.ftl
@@ -1,0 +1,10 @@
+﻿trait-category-mood = Настроение
+
+trait-fortitude-name = Моральная стойкость
+trait-fortitude-desc = Вы не такой эмоциональный человек как остальные, вам плевать на все трудности и радости жизни.
+
+trait-hypomania-name = Гипомания
+trait-hypomania-desc = Обычно вам чертвоски весело, но любой негатив вас сильно расстраивает!
+
+trait-depression-name = Депрессия
+trait-depression-desc = Чувак, эта вечеринка отстой. Я, блин, ненавижу этих людей.
