+++ Resources/Prototypes/_Backmen/Mood/genericNegativeEffects.yml
@@ -42,6 +42,10 @@
   moodChange: -20
 
 - type: moodEffect
+  id: TraitFortitude
+  moodChange: -10
+
+- type: moodEffect
   id: Dead
   moodChange: -1000
 
@@ -59,5 +63,5 @@
 
 - type: moodEffect
   id: SurgeryPain
-  moodChange: -15
-  timeout: 180
+  moodChange: -30
+  timeout: 90
