+++ Resources/Prototypes/_Backmen/Traits/categories.yml
@@ -1,0 +1,4 @@
+- type: traitCategory
+  id: MoodTraits
+  name: trait-category-mood
+  maxTraitPoints: 1
