+++ Resources/Prototypes/_Backmen/Traits/disabilities.yml
@@ -1,4 +1,3 @@
-
 - type: trait
   id: Monolingual
   name: trait-monolingual-name
